add: docker のテストを追加

+++ build_util/check_release_build.py
@@ -1,12 +1,8 @@
-"""
-ビルド結果をテストする
-"""
 import argparse
 import json
 import time
 from io import BytesIO
 from pathlib import Path
-from subprocess import Popen
 from urllib.parse import urlencode
 from urllib.request import Request, urlopen
 
@@ -15,14 +11,16 @@ import soundfile
 base_url = "http://localhost:50021/"
 
 
-def test_release_build(dist_dir: Path) -> None:
+def test_release_build(dist_dir: Path, skip_process: bool) -> None:
     run_file = dist_dir / "run"
     if not run_file.exists():
         run_file = dist_dir / "run.exe"
 
     # 起動
-    process = Popen([run_file.absolute()], cwd=dist_dir)
-    time.sleep(120)  # 待機
+    process = None
+    if not skip_process:
+        process = Popen([run_file.absolute()], cwd=dist_dir)
+        time.sleep(60)  # 待機
 
     # バージョン取得テスト
     req = Request(base_url + "version")
@@ -52,15 +50,17 @@ def test_release_build(dist_dir: Path) -> None:
         manifest = json.loads(res.read().decode("utf-8"))
         assert "uuid" in manifest
 
-    # プロセスが稼働中であることを確認
-    assert process.poll() is None
+    if not skip_process:
+        # プロセスが稼働中であることを確認
+        assert process.poll() is None
 
-    # 停止
-    process.terminate()
+        # 停止
+        process.terminate()
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--dist_dir", type=Path, default=Path("dist/"))
+    parser.add_argument("--skip_process", action="store_true")
     args = parser.parse_args()
-    test_release_build(dist_dir=args.dist_dir)
+    test_release_build(dist_dir=args.dist_dir, skip_process=args.skip_process)
